+++ JavaScript/JsExampleLibrary.kt
@@ -1,0 +1,5 @@
+package net.milosvasic.fundamental.kotlin.javascript
+
+fun helloJS() {
+    println("Hello from JavaScript!")
+}